Split out geo_ip and OwnedHostCall from js-compute-builtins.cpp
